Successfully subscribed to store in main related items and outfits component
