--- client/src/components/Comparing.jsx
@@ -1,28 +1,0 @@
-import React from 'react';
-
-function Comparing () {
-  return (
-    <div className="comparing-container">
-      <h6>Comparing</h6>
-      <table className="comparing-table">
-        <tr>
-          <th>current product</th>
-          <th></th>
-          <th>related product</th>
-        </tr>
-        <tr>
-          <td>check</td>
-          <td>uv blocking</td>
-          <td></td>
-        </tr>
-        <tr>
-          <td>brass</td>
-          <td>buttons</td>
-          <td>plastic</td>
-        </tr>
-      </table>
-    </div>
-  );
-}
-
-export default Comparing;
