--- client/src/components/RelatedAndOutfits.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import RelatedItems from './RelatedItems';
-import Outfits from './Outfits';
-import Comparing from './Comparing';
-
-function RelatedAndOutfits () {
-
-  return (
-    <div>
-      <RelatedItems />
-      <Outfits />
-      <Comparing />
-    </div>
-  );
-}
-
-export default RelatedAndOutfits;
