+++ client/src/components/relatedItems/Comparing.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+
+function Comparing () {
+  return (
+    <div className="comparing-container">
+      <h6>Comparing</h6>
+      <table className="comparing-table">
+        <thead>
+          <tr>
+            <th>current product</th>
+            <th></th>
+            <th>related product</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td>check</td>
+            <td>uv blocking</td>
+            <td></td>
+          </tr>
+          <tr>
+            <td>brass</td>
+            <td>buttons</td>
+            <td>plastic</td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  );
+}
+
+export default Comparing;
