+++ package.json
@@ -58,6 +58,7 @@
     "axios": "^0.21.1",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "lodash": "^4.17.21",
     "morgan": "^1.10.0",
     "react": "16.14.0",
     "react-dom": "16.14.0",
